Workaround for OPAM env issues
Chad was the first to notice that ocamlfind wasn't being found sometimes.
It turns out that this stems from how OPAM works with the user's environment.
As a workaround from this (which, to me, is a design flaw in OPAM) we have
added a line in the vagrant user's .profile to run the necessary incantation:

eval `opam config env`

whenever openining a new shell session.