Change 0 sample output of prob and fix bench.hs
The output of prob when using 0 samples is now identical to when there is
sampling (even though the information isn't 'new'). This means that we can
parse the output of prob uniformly for all experimental runs.

I also changed the Result type to contain Integers instead of Ints in order to
avoid overflow.