Thread sample point through _all_ queries.
I had messed up the sampling procedure and I was only running
the most recent query on a sample point. This commit fixes that
by threading all of the queries to each region in a list
of point -> bool functions. We test each query and short-circuit
the computation if one returns false (this is done with the new
function list_func_and in util.ml)