Remove check for dsa that was breaking records
This is mostly a whitespace diff, the sole actual change is removing the check
for the dsa CLI flag that was preventing records from working (line 111 of
prob.ml).

If we ever want to use dsa in the future we will have to fully execute the
query_stmt _before_ the call to sa_of_stmt on line 112 of prob.ml

Additionally I've converted one of Stephen's query files to use records. I
tested it against the original version and the results are the same.