This commit addresses TAM-61 and TAM-62
TAM-61 is about cleaning up the output. Now the default
output is very minimal with very short summaries of each
query being analyzed.

TAM-62 is about a outputting the _real_ revised belief at the
end of the computation. It had to do with how Piotr was handling
the possibility of a policy rejected the query. We now short
circuit that if we aren't using a verbose output.