Piotr left-nests SSeqs! And that's okay!
This shouldn't have thrown me off as much as it did but I was trying to find a
bug and had to output the intermediate ASTs, they were confusing me and so I
kept outputting more information. I formatted the below by hand. Luckily it, it
did help me solve the bug in the end, but it's not that left-nesting is somehow
bad, I was just expecting SSeqs to be like a standard cons list and it threw me
for a loop.

The issue is that the leftmost statment in the branches of an SIf _is not_
the succ stmt like it is for the rest of the program.

Seq (Seq (Seq (int32 result; result = 0);
          Seq ( int32 test; test = 0)
         );
      if (ship == 1) then
      Seq (( if ((ship1_loc <= 200) and (ship1_loc >= -200)) then
                result = test
              else
                skip
              endif;)
              test = 2
           )
      else
        skip
      endif)

-----------------
The above, flipped
-----------------

Seq ((if (ship == 1) then
      Seq ( test = 2;
              if ((ship1_loc <= 200) and (ship1_loc >= -200)) then
                result = test
              else
                skip
              endif)
        else
          skip
        endif;)
  Seq (Seq ( test = 0; int32 test);
       Seq ( result = 0; int32 result)
      )
)