Remove splitter type/make_splitter, etc.
This changeset removes the exposed splitter type in the pstateset API, and the
corresponding functions make_splitter, split_many_with_splitter. It isn't
entirely clear why they are part of the public interface.

With this change, and the subsequent implementation of splitter logic in the
dpstatesetmaker, decomposition should be functionally complete.